chore: improve code structure by moving touched pages into a separate file
